Changing name and adding error tracking
